requirements.txt
facebook-scraper
mysql-connector-python~=8.0.25
requests~=2.25.1